[refactor]:refactor tidb parser files
1. move jflex and cup to src/main/resources/tidb
2. refactor parser output dir to  ecnu.db.tidb.parser
