+++ src/main/java/ecnu/db/tidb/TidbAnalyzer.java
@@ -7,8 +7,8 @@ import ecnu.db.analyzer.online.AbstractAnalyzer;
 import ecnu.db.analyzer.online.node.ExecutionNode;
 import ecnu.db.analyzer.online.node.ExecutionNode.ExecutionNodeType;
 import ecnu.db.analyzer.online.node.RawNode;
-import ecnu.db.analyzer.online.select.tidb.TidbSelectOperatorInfoLexer;
-import ecnu.db.analyzer.online.select.tidb.TidbSelectOperatorInfoParser;
+import ecnu.db.tidb.parser.TidbSelectOperatorInfoLexer;
+import ecnu.db.tidb.parser.TidbSelectOperatorInfoParser;
 import ecnu.db.constraintchain.filter.SelectResult;
 import ecnu.db.dbconnector.DatabaseConnectorInterface;
 import ecnu.db.schema.Schema;
